chore(activate): Move _FLOX_WATCHDOG_BIN to getopt
To prevent it leaking from another activation when `activate` is called
directly. We should adopt this same pattern for other environment
variables that are only used by a single invocation of `activate` (e.g.
not `_FLOX_ACTIVE_ENVIRONMENTS`) but this is the most pressing one for
now.

I considered naming it `--watchdog-bin` but it seems like an unnecessary
mouthful when the error describes what it means and it's also the
conditional about whether the watchdog is started.

It's worth noting that the default value of an empty string before
parsing options will guard against any existing values from activations
with previous versions of FLOX.

This has been causing all of the "can be run directly" tests to fail for
me locally after unsetting `_FLOX_ENV_*` in b2f6d777 but I've been
partly ignoring it because I tend to run the tests focused:

    ✗ zsh: activation script can be run directly with --turbo [296]
       tags: activate activate
       (from function `assert_success' in file /nix/store/h7inz07x6739s38i02532g5pj6dlnx8k-bats-with-libraries-1.11.0/share/bats/bats-assert/src/assert_success.bash, line 42,
        in test file activate.bats, line 782)
         `assert_success' failed
       …

       -- command failed --
       status : 1
       output : /nix/store/k44nfncj0mz3ncr1l97bhfl9qh0k4srj-flox-activation-scripts/activate.d/start.bash: line 67: _FLOX_ENV_LOG_DIR: unbound variable
       --