test(containerize): Disable Rosetta by default
Running `podman machine start` brings up a prompt to install Rosetta if
it's not already installed. I don't want to install it on my laptop and
believe this also the cause of the test failures for Darwin in CI on
GitHub Runners:

    # Machine init complete
    # To start your machine run:
    #
    # 	podman machine start
    #
    # Starting machine "podman-machine-default"
    # Error: vfkit exited unexpectedly with exit code 1
    # Machine "podman-machine-default" stopped successfully

This is not disabled when `$NIX_SYSTEM` is x86 because we use Rosetta to
perform x86 builds on aarch64 hosts with our remote builders. Yannik
verified locally that the value should be correct from the closure:

> `nix run nixpkgs#legacyPackages.x86_64-darwin.nix -- eval --expr 'builtins.currentSystem' --impure` returns "x86_64-darwin" when run on an aarch64 machine, so assuming that everything uses absolute references into its closure,
> `flox/flox#packages.x86_64-darwin.flox-cli-tests` should also run with a `x86_64-darwin` nix and return the "right" system