fix: export internal variables required for in-place activation
We were exporting required flox internal variables when first activating
an environment but not when attaching, and that caused problems for
activation scripts that were expecting those variables to be defined.
This patch moves those activation blocks further down where they can
be defined for all activation types and makes them consistent across the
bash, fish, and tcsh cases. The zsh case is handled differently on account
of the way that zdotdir works, and is not a problem.