refactor(init): Remove get_default_package_if_..
Renamed `try_find_compatible_version` to `try_find_compatible_package`
and made the `version` optional. It otherwise has exactly the same
functionality as `get_default_package_if_compatible` because the
`default` package group has no special meaning after we switched from
pkgdb to catalog for resolution.

It does drop the version field from the debug trace when a package can't
be found, otherewise we have to clone it earlier on, but I don't think
it adds much value when we already have it in the first debug trace of
the function.