refactor(init): Remove get_default_package
Replace `get_default_package` with a thin shim over the recently
modified `try_find_compatible_package` which surfaces the same errors
with the addition of the word "compatible" to make it clear that the
package isn't missing entirely.

`default` has no special meaning after we switched from pkgdb to catalog
for resolution.