feat(upgrade): add `--dry-run` option
With `--dry-run` upgrade will use `Environment::dry_upgrade` to produce an upgrade result,
without changing the current environment.
If anything could be applied, the result is then rendered back to the user as

    - <install id>: <old version> [-> <new version]

Note, that upgrades not affecting the current system are omitted,
or assumed to be the same as for the current system.
If upgrades, _only_ affect different systems, the message will read

    Upgrades are not available for 'name' on this system, but upgrades are
    available for other systems supported by this environment.

This commit also touches up the wording of related upgrade messages