refactor: rework podman vm lifecycle
The original intention for this commit was to cache the VM between test
runs so that we don't need to repeatedly download the VM image, create
the VM, and then boot it from scratch for every test run. That turned out
to have complexities around race conditions between multiple CI jobs on
a single runner, VM lifecycle, etc.

Instead, this commit refactors the VM setup logic to make it clearer to
understand how the temporary directories are structured. It also strikes
a middle ground of convenience for local development and CI by creating
a new VM for each test run in CI and requiring that developers start the
VM themselves during local development.

This requires that developers create a VM called `flox-containerize-vm`
with certain directories mounted. This is documented in CONTRIBUTING.md,
and essentially instructs the user to create a VM using the same command
that's in `containerize.bats`. Developers must also make this VM the
default during development. This VM will also persist data outside the
test suite in the host machine's XDG_DATA_HOME and XDG_CONFIG_HOME
directories.