test(init): Add tests for Node::new
It holds a lot of logic that isn't covered by the unit tests of the
child functions and doesn't currently feel very safe to modify or
refactor.

There's some overlap here with the unit tests for the child functions
but I think there's still value in keeping both because they will give
more specific errors vs wider coverage if we make any regressions.

It's worth noting that these new tests don't cover all of the cases
where resolution might affect the outcome because they are numerous and
they are heavily influenced by what the catalog returns.