refactor(commands): {config->flox}.features
Use `flox.features` instead of `config.features` for all of the build
and publish commands which are feature flagged.

Unfortunately we have to carry two copies of it around per 7ba717f6 but
`Flox.features` is easier to access because defaults have already been
applied and we don't need to move/ref it in doing so.