feat(publish): Merge cache opts from config
Allow `publish.store_url` and `publish.signing_key` to be provided from
config. Any CLI arguments that are provided are merged over the top of
these. The names match the (renamed) arguments.

The URL and key can be mixed-and-matched from different sources but they
are required to be mutually present or absent. For this reason the
individual entries are now `Option<>`. Related discussion:

- https://flox-dev.slack.com/archives/C07GRNPFNJE/p1737037013075899

We may in the future decide that it doesn't make sense to provide these
from host config because you would only change them on a per-project
rather than per-host basis but we want to make it easier to use and
gather some feedback in the meantime. Related discussion:

- https://flox-dev.slack.com/archives/C07GRNPFNJE/p1737038217746299

These new keys haven't been added to the `flox-config(1)` docs because
we're not currently publicising the feature.