feat(containerize): PascalCase -> kebab-case
Rename `containerize.config` keys from PascalCase to kebab-case in the
user-facing manifest and only convert them when passing over to the
provider. Rationale from Jenny's update to the DR:

- https://www.notion.so/floxdev/011-Containerize-image-options-14a744c8e31680f09425d84de06a3bd4?pvs=4#14a744c8e3168199a9d0e320587a089d

> We will change the capitalization of the keys to kebab-case instead of
> PascalCase. This will be consistent with the rest of the Flox manifest
> fields. Since the values for each key will have a slightly
> different (nicer) format in the manifest.toml vs the spec, we don’t
> think it’s necessary to follow the capitalization convention of the spec
> file.

We'll maintain two structs, one user facing and the other provider
facing, because we're also going to modify the behaviour of values that
are unnecessary maps in a subsequent commit.

We considered using the third-party `oci_spec::image::Config` but we
don't need any of the other functionality so it doesn't seem worth
carrying an extra dependency.

This isn't a breaking change because it hasn't been released yet.