feat(containerize): Simplify arrays
Present `exposed-ports` and `volumes` as `Vec<String>` instead of
`BTreeMap<String, BTreeMap<(), ()>>>` in the manifest. Perform an
internal conversion for Go's `map[string]struct{}` on the way over to
the provider.

This will be less surprising to users, especially if they aren't
familiar with the internals of OCI and Go. It also makes it much easier
to describe the syntax in our docs.

This isn't a breaking change because it hasn't been released yet.