refactor(containerize): Vec -> BTreeSet
Change `exposed_ports` and `volumes` in the manifest to use
`BTreeSet<String>` instead of `Vec<String>` to better reflect that they
get de-duplicated on the way over to the OCI format.

This doesn't make any functional difference because TOML doesn't have
any notion of a set and they were already de-duplicated in the
conversion to a `BTreeMap` but it's technically correct and marginally
easier to reason about.