feat(containerize): print name of container image
Before we were piggy backing on the output of `docker load`
to communicate the name of the image to the user, specifically when loading into a runtime.
Since docker logs are now hidden by default, that message is not visible anymore,
so we include it in our own success message.