refactor: Extract known_store_path test helper
I plan to make `NIX_BIN` private so that we can control a default set of
feature arguments but there are a number of places where we also use it
to get a known store path for tests. Refactor these to use a test helper
rather than accessing the static or environment variable directly.

Also incldues a missing `#[cfg(test)]` because `commands::upload` was
complaining about unused imports on build after adding the helper.