refactor: Add providers:nix::nix_base_command
To ensure that we apply the correct set of default options to all `nix`
commands. Not doing so causes `flox publish` to fail if the system
defaults don't match:

    % NIX_CONFIG='experimental-features = ""' ft publish hello2
    ❌ ERROR: Unable to interpret derivation info: EOF while parsing a value at line 1 column 0: EOF while parsing a value at line 1 column 0

This now works:

    % NIX_CONFIG='experimental-features = ""' ft publish hello2
    ✅ Package published successfully.

    Use 'flox install dcarley/hello2' to install it.

We also had the same issue with `flox containerize` in 0aeeb92a. This
wouldn't have helped for `ContainerProxy` though, where the commands are
nested and we have to apply additional options like aecd753f.

I've standardised on `--optional extra-experimental-features` rather
than `--extra-experimental-features` per 108f2825.