feat(containerize): print full error of failing subcommands
The stderr content of subcommands called by `containerize` are now logged to tracing
with `info` verbosity.
That means by default, users will not see these logs, including any errors that occur.
This commit will in case of errors, take the _collected_ content and add it to the error,
that is eventually shown to the user.