fix(tests): round timestamps down to seconds
We saw a few tests fail sporadically, due to ser/de of iso8601 timestamps
created by `time::serde::iso8601` [1].
Deseralization of timestamps, casued the resulting native value
to be off by -1 nanosecond.
The underlying issue seems to be the storage of time as a float
with tricky behaviour at this granularity.
We discussed dropping `time`
(which was only first introduced to `flox-rust-sdk` for this task)
in favor of using `chrono`,
but for the time being decided to fix the tests instead,
done here by lowering the granularity to seconds,
which should avoid the float floor'ing.