test(activate): Add timeout to activate_finished
Prevent the tests from hanging if we make a breaking change to activate
which causes the `echo > activate_finished` command to never execute.

For cases where we're redirecting the output you may need to inspect the
output file to debug the problem further. We're not consistently doing
the same thing with the output though, so trying to be too clever in a
helper function might not have the desired effect, and the same problem
should be apparent in other simpler tests anyway.