feat(containerize): add single-user install capability
flox containerize expected to always run as root inside the container, since the
Nix store was owned by root, and the following files were basically missing.

    /etc/passwd
    /etc/groups
    /etc/shadow

This commit changes that by extending the container with a "fakeNss" to provide
those missing files, and also abuse the fact that we can insert custom lines
during build-time. We derive most information from containerize.config.user
inside Flox's toml.

All that being said, it allows us to do the following.

    $ flox containerize --file='-' | docker load
    ✨ 'tmp.GtqsVDTJut:latest' written to stdout
    Loaded image: tmp.gtqsvdtjut:latest
    $ # Works perfeclty fine when we just run it with the default user root
    $ docker run --rm -it tmp.gtqsvdtjut:latest 'cowsay hello'
     _______
    < hello >
     -------
            \   ^__^
             \  (oo)\_______
                (__)\       )\/\
                    ||----w |
                    ||     ||
    $ # There is a "bryan" user on this system, let's try to use that...
    $ docker run --user='bryan' --rm -it tmp.gtqsvdtjut:latest 'cowsay hello'
    Error: failed to acquire lockfile

    Caused by:
        Permission denied (os error 13)
    $ # This is expected since the Docker image wasn't build with extra users in
    $ # mind. Let's try changing that with the changes in this commit...
    $ echo '[containerize.config]' >> .flox/env/manifest.toml
    $ echo "user = 'bryan:bryan'" >> .flox/env/manifest.toml
    $ flox containerize --file='-' | docker load
    ✨ 'tmp.Fc9KpHvftX:latest' written to stdout
    Loaded image: tmp.fc9kphvftx:latest
    $ # root still works...
    $ docker run --rm -it tmp.fc9kphvftx:latest 'cowsay hello'
     _______
    < hello >
     -------
            \   ^__^
             \  (oo)\_______
                (__)\       )\/\
                    ||----w |
                    ||     ||
    $ # But this now also works...
    $ docker run \
    >  --user='bryan:bryan' \
    >  --rm -it \
    >  tmp.fc9kphvftx:latest 'cowsay hello'
     _______
    < hello >
     -------
            \   ^__^
             \  (oo)\_______
                (__)\       )\/\
                    ||----w |
                    ||     ||

Cool, now we can use non-root user inside of the container. Note that the UID
and GID are currently hard-coded to 501, since I'm not sure the OCI spec can
support setting both at the same time. It looks to me as if its and either or
thing.

Nixpkgs' streamLayeredImage now also uses the correct architecture when building
containers. We're putting binaries coming from the containerPkgs attribute into
the final  image, so the container's architecture config should be set to
containerPkgs' architecture as well. Using containerPkgs.go.GOARCH is just a
shorthand to give us the correct format that the OCI spec is expecting.

I also took a little time to refactor the mkContainer.nix file, the biggest
change being that I moved some of the attributes that were first inside of a let
binding, to the function's argument attribute set. This just made it a bit
easier for me to itterate on this issue, without having to constantly rebuild
Flox.

Ref: flox/product#822