feat: return build metadata in BUILD_RESULT_PIPE
This patch updates the package builder to return the build data provided by
`nix build --json`, enhanced with `pname` and `version` as set in the manifest.
In the case of `version` specifically, this patch also adds logic to evaluate
the "file" and "command" modes at the point of performing a build.