test(activate): inputrc disable bracketed paste
To remove the `2004h` and `2004l` control codes in `expect` output so
that we can match on complete output:

    % FLOX_SHELL=bash expect cli/tests/activate/attach.exp ~/tmp "true" | cat -ev
    spawn /Users/dcarley/projects/flox/flox/cli/target/debug/flox activate --dir /Users/dcarley/tmp^M$
    M-bM-^\M-^E Attached to existing activation of environment 'tmp'^M$
    To stop using this environment, type 'exit'^M$
    ^M$
    true && exit^M$
    ^[[?2004hflox [tmp shell] bash-5.2$ true && exit^M$
    ^[[?2004l^Mexit^M$

We'll need a different approach for zsh and fish but I'm going to defer
that for now.

I also experimented with using `TERM=dumb` which works in most cases but
causes fish to repeat the shell prompt many many times.