fix: revert to previous target structure keeping result link in path
Previous makefile changes removed the build's primary result link target
from the dependency graph, but I forgot that was necessary for checking
for failed pure builds.

This patch reverts much of the previous patch and introduces support for
the JSON build metadata as a downstream requisite of the result target,
fixes a few other bugs in the previous patch and restores the original
usage target to minimize the diff from origin/main.

It also orders the make invocation as the more typical `make [targets] [vars]`
ordering rather than interspersing build targets among variable overrides
and stops passing the BUILDTIME_NIXPKGS_URL variable not required for the
clean target(s).