refactor(manifest): skip_serializing_{if,none}
Use the `skip_serializing_none` macro on structs rather than
`skip_serializing_if` on individual fields, for consistency with the
other structs in the manifest.

Arguably it's a little less obvious and could be missed when writing a
new child struct but we could equally forget to add it to new fields and
consistency is better.