refactor(catalog): make `extra_headers` config, non-optional
We can omit a layer of wrapping `CatalogClientConfig::extra_headers` in an `Option`,
because the field is a `BTreeMap`.
`BTreeMap` implements `Default` by returning an empty map,
which wrt the client behaves equivalently to `None`.
I.e. there is no difference between `None` and `Some(BTreeMap::default())`.
Besides, it makes modifying options more ergonomic,
as it removes the need to check whether an option was set.