feat(catalog): add an enum for Quality of Service classes
`CatalogQoS` is represents the QoS classes exposed by macos
and described for example in [1].
As of now we don't have a usecase for the majority of these,
and will start with only using `background` for `upgrade`/`resolve`
requests made by the `check_for_upgrade` command,
which runs ... in the background.
Additional classes are included to provide a "proven" framework
for QoS classes in full.

[1]: <https://blog.xoria.org/macos-tips-threading/>

fix: provide display name for each class manually

`display(rename_all = "snake_case")` does not actually seem to be a thing,
that is supported by `derive_more` [1].
Here, for sake of simplicity, i manually named the fields.
Alternatively, we could change the "extra_headers" to `Map<String, serde_json::Value>`,
which would then allow using the `Serialize` derive here as intended.

[1]: <https://github.com/JelteF/derive_more/issues/216>