feat(check_for_upgrade): set QoS class header to `background`
Upgrade checks running in the background,
may eventually be handled differently,
e.g. given lower priority over user attended resolution requests.
One way of doing so, is by attaching a QoS class header to requests
made by  the `check-for-upgrades` command,
that describes the requested Quality of Service.
In this case we accept a low priority/quality as the request isn't user attended.

@billlevine brought up that if we start mutating the catalog client
or `Flox` object in multiple places, it would be preferable
to do som in a more scoped way and have the changes be reverted at some point [1].
For now, in this command we're modifying the `Flox` object only once
and for the rest of the command's (short) lifetime.
A possible future improvement was sketched out in the comment above [1].

[1]: <https://github.com/flox/flox/pull/2658#discussion_r1932362747>