refactor: amend flox version detection and document
The version is determined by the following rules:
1. `github:flox/flox#flox`, provides a wrapper that sets `FLOX_VERSION`.
   This is the main production artifact and canonical version.
2. Our `just` targets will set `FLOX_VERSION` using the current git tag,
   so `just` builds will have the correct updated version _with_ git metadata.
3. `cargo build` when run outside of `just` will fallback to `0.0.0-dirty`.
   This is the version that also local IDEs / rust-analyzer will use.
   However, binaries built this way may fail to run in some cases,
   e.g. `containerize` on macos which relies on the flox version.

In short, if your work requires the current version of flox,
use a nix or `just` build artifact.
If you use flox _inside_ of another flox activation, call it with `just flox`
or use the wrapped nix build.

build: drop '-dirty' suffix from `just` set FLOX_VERSION

Set the FLOX_VERSION variable so that it can be used in the build/runtime
It's important to add the git revision to the version string,
so to that `containerize` can build the correct version of flox in CI.
While technically we'd want to add `-dirty` to the version string if the
working directory is dirty, we omit this here because in practice
it causes tests to fail that expect a FLAKE_VERSION to be "clean",
and doesn't add practical information.