feat: add Rust-based path helper commands
This adds two new commands to `flox-activations`: `set-env-dirs` and
`fix-paths`.

The `set-env-dirs` command handles setting `FLOX_ENV_DIRS`, which the
awk-based path helper had subsumed into it and removed from the CLI.
This command prints sourceable output that exports the new value of
`FLOX_ENV_DIRS`.

The `fix-paths` command handles prepending the `/bin` and `/sbin`
subdirectories of each environment to `PATH`. It similarly handles the
`/share/man` subdirectory for `MANPATH`.