refactor: don't require flag for all commands
The way that the arguments are constructed at the moment requires that
every single command of `flox-activations` pass in a `--runtime-dir`
flag. That may make sense for the existing commands, but it's not
necessary for the new commands as they only operate on the data passed
in as arguments.

This change makes the `--runtime-dir` argument required only for those
commands that need it.