test: possible fix for expect flakey test
There was a flakey test in
https://github.com/flox/flox/actions/runs/13058298843/job/36434896986#step:5:344
where it appears expect sent "true && exit" before "To stop using this
environment, type 'exit'" was printed.

Add an expect for the prompt so the assert_output can depend on a
reliable ordering.

If using `assert_output` continues to cause flakey tests, it may be better
to use `--partial`, but we can start by seeing if this fix works.

Add an expect for the prompt throughout our expect files to ensure a
reliable ordering throughout.