refactor(manifest): Split raw and typed modules
`models::manifest` was getting pretty large at 2600~ LoC including tests
and we're about to add support for composite manifests which will make
it larger, or live separately, so split it into "raw" (for writing) and
"typed" (for reading) sub-modules.

It wasn't feasible to break this into any smaller git commits but most
clients should represent it as a rename from `manifest.rs` to `raw.rs`
with modifications to the stuff that's stayed and deletions for the
stuff that's moved to `typed.rs`.

The only significant changes are:

- `pub(super)` -> `pub(crate)` visibilty
- `RawManifestError` split from `ManifestError`