refactor: split out retry for valid systems
This factors out the second installation attempt from the error handler
so that the error handler can be smaller and more focused. There was
too much logic in the error handler, and it made sense to keep the
retry nearer to the first installation attempt.

The other motivation for this change is that we're going to group the
successful installations together under a single checkbox, and the
implementation for that is cleaner with this refactoring.