test: skip compatability check for relase commits
We recently added a test to check activation compatability with the
previous release. As the test depends on the latest flox version having
been released on GitHub, the test fails when run on a commit that bumps
flox version but has not yet actually been released.

We could alternatively make sure to tag the release commit and then
re-run CI, but that seems like it would introduce unnecessary friction
into the release process.