fix: daemonize invocations of `nix store delete`
The `nix-daemon` can be slow to respond to `nix store delete` invocations
for a variety of reasons, and users should not have to wait for these to
complete, particularly as the success of this command cannot be guaranteed
as users can create alternate GC roots to the packages being deleted.

This patch eliminates the slowness by invoking the deletion in the
background by way of the `daemonize` wrapper.

Closes #2670.