fix: don't prompt for default env in current dir
When deciding which environment to perform an operation on, we already
give precedence to the current directory over an activated default
environment.

Add a similar exception to give precedence to an activated environment
over a default environment in the current directory.

This avoids the annoyance of a prompt in the situation, where the user
most likely wants to operate on the active environment
- Start shell in ~ (which contains an activated default environment)
- flox activate -r owner/name
- Run a flox environment command

Notably, this does not regress the behavior that flox environment
commands default to an activated default environment when no other
environment is active, which makes Flox feel more similar to a system
package manager.