refactor: add IntoInner trait
Several of the fields in the manifest are simply wrappers around
collections. When implementing merges you'll need to access those inner
collections. This is a uniform interface for doing so.

This was done as a macro simply because the actual source code of the
implementation looks *exactly* the same for all of these collection
wrappers.

This also removes the `derive_more::Deref` derives because using Deref
for type conversion is an antipattern called "deref polymorphism". See
https://rust-unofficial.github.io/patterns/anti_patterns/deref.html#discussion
for a discussion of why this is an antipattern.

Finally, this trait makes it explicit when you're accessing the wrapped
collection, which was previously implicit.