fix(activate): Prevent dupe upgrade notifications
Don't print the upgrade notification if the environment is already
active, which can easily happen if you follow our default environment
instructions for Bash or Zsh which do an in-place activation from both
an RC and profile file, e.g.

Before:

    % bash
    bash-5.2$ eval "$(flox activate -d ~/tmp)"
    ℹ️  Upgrades are available for packages in 'tmp'.
    Use 'flox upgrade' to get the latest.

    flox [tmp term] bash-5.2$ eval "$(flox activate -d ~/tmp)"
    ℹ️  Upgrades are available for packages in 'tmp'.
    Use 'flox upgrade' to get the latest.

    flox [tmp term] bash-5.2$

After:

    % bash
    bash-5.2$ eval "$(flox activate -d ~/tmp)"
    ℹ️  Upgrades are available for packages in 'tmp'.
    Use 'flox upgrade' to get the latest.

    flox [tmp term] bash-5.2$ eval "$(flox activate -d ~/tmp)"
    flox [tmp term] bash-5.2$

This partially duplicates some of the existing activation check in
`Activate::activate` but it's cheap to execute, it's clearer to
co-locate with the other conditionals for skipping the notification, and
it makes it easier to unit test.