test: cleanup background activations
When a backgrounded activation reached timeout, we would fail the test
suite, but we wouldn't clean up the activation or watchdog left behind. This
could lead to build up of orphaned processes.

To reproduce this, I added `sleep 1s` before `echo > activate_finished`
and lowered the wait_for_background_activation timeout to .5s. Then
running:
flox-cli-tests activate.bats -- -f "interactive: bash attachs to an activation from the previous release"; pstree -ws bats

Would leave behind:
```
-+= 00001 root (launchd)
 |--= 54832 matthew /nix/store/xgiggnk7c2d1n0xy8x2cyvml42gjngc4-flox-watchdog-1.3.12/libexec/flox-watchdog --disable-metrics --log-dir /private/var/folders/2w/r9qm37r12qs4xn3sjn4cv2ym0000gn/T/nix-shell.27hGkZ/bats-run-pXupvJ/test/1/project-1/.flox/log --socket /tmp/flox.tests.9Mv4kw/tmp.3CsewV5DHc/run/flox.87161cb9.sock --flox-env /private/var/folders/2w/r9qm37r12qs4xn3sjn4cv2ym0000gn/T/nix-shell.27hGkZ/bats-run-pXupvJ/test/1/project-1/.flox/run/aarch64-darwin.project-1.dev --activation-id dd9d3c20 --runtime-dir /tmp/flox.tests.9Mv4kw/tmp.3CsewV5DHc/run
 \-+- 54777 matthew /nix/store/jin4ifpw4bvi554g02phy668gvaylqn1-bash-interactive-5.2p32/bin/bash --noprofile --norc -s
   \--- 54847 matthew /nix/store/jin4ifpw4bvi554g02phy668gvaylqn1-bash-interactive-5.2p32/bin/bash -c sleep 1s && echo > activate_finished && echo > /var/folders/2w/r9qm37r12qs4xn3sjn4cv2ym0000gn/T/nix-shell.27hGkZ/bats-run-pXupvJ/test/1/project-1/teardown_activate
```

To fix this, start the background activation in a new process group, and
kill the process group if the timeout is reached.

A new process group had to be used since Bash doesn't cleanup it's
children when it receives SIGTERM.