feat(watchdog): Loop on GCing all log types
Rather than special casing the watchdog logs.

It's possible for `process-compose` to be stopped and started multiple
times for a long running watchdog, so long as there is at least one
attachment to the activation. And we're about to add handling for
`check-upgrade` logs.

So we should attempt to garbage collect for the life of the watchdog.
This is done every 1hr which still seems reasonable.