refactor(core): Extract upgrade-check log pattern
Share the construction of the filename pattern between the CLI and
watchdog, so that they don't get out-of-sync and evade GC.

This can't be constant because `format!` needs a string literal.

We can't do the same for services logs because the filename is
constructed in bash. And newer watchdog logs use a different pattern for
log rotation.