fix(tests): don't use random podman-user as USER when running podman
flox/flox@a9299e5 introduced the idea that we had to set $USER to something to
avoid some errors when running podman in rootless mode (which is its default
mode). flox/flox@05a08cb actually set it to `podman-user`, which is what I saw
in some errors.

The errors mentioned in flox/flox@a9299e5:

    Error: command required for rootless mode with multiple IDs: exec: "newuidmap": executable file not found in $PATH

and

    time="2024-11-06T19:12:54Z" level=error msg="running `/nix/store/xi2asrhj9m8vzcg4zmp03947vsjd25x2-shadow-4.16.0/bin/newuidmap 2393938 0 1015 1 1 689824 65536`: newuidmap: write to uid_map failed: Operation not permitted\n"
    Error: cannot set up namespace using "/nix/store/xi2asrhj9m8vzcg4zmp03947vsjd25x2-shadow-4.16.0/bin/newuidmap": should have setuid or have filecaps setuid: exit status 1

Are both solved by these changes. One of them by adding /run/warppers/bin to
PATH, since NixOS is weird by how setuid binaries work on the system and puts
the ones that will work outside of the Nix store under /run/wrapper. And the
other by actually including it in the PATH to begin with in the test suite.
Because we're very paranoid about adding stuff to PATH apparently...

I'm not sure what the implications of these changes are on macOS, but if I'd
have to guess, they probably wont affect them.

Refs: flox/flox@a9299e5514bda59fb1a82c05ec2a05ef201ca1b0
Refs: flox/flox@05a08cbf7cf4c99cd351355f995a316fcecfaab3