test: avoid triggereing a rebuild over fake env
Flox will now rebuild environments if the rendered lockfile does not equal the definition `.flox/env`.
in the bats test `can use bundled interpreter to mitigate broken bundled interpreter`
we create a custom "fake" rendered path in `.flox/run` which did not contain a lockfile.
It is thus being rebuilt, which fails on account of it not pointing to a store path,
and would further fail regardless because the properties of the fake environment are removed.
Instead make sure to copy over the lockfile of the env definition to ensure flox wont rebuild the environment.