test(mkContainer): add unit tets of internal functions
This adds some unit tests to the internals of mkContainer.nix, especially for
the part that parses the uname:gname from containerize.config.User. The tests
can be run as follows.

    $ nix eval --impure --expr '(import ./mkContainer/mkContainer.nix { nixpkgsFlakeRef = "github:nixos/nixpkgs?ref=nixos-24.11"; environmentOutPath = null; system = builtins.currentSystem; containerSystem = builtins.currentSystem; }).passthru.tests'
    $ [ ]

If the above command returns _anything_ else then [ ], it has failed. The items
in that lists are the tests that have failed. So, and empty list means all tests
have passed.

I wasn't quite sure where to put these tests to be run by CI, so I've just added
a comment on the passthru section in mkContainer/mkContainer.nix for the time
being on how to run these tests. There's also probably a better way to pass
nixpkgsFlakeRef to the function, but I can't be bothered figuring that out right
now.