build(interpreter): split source files and rename package
Split source files of the environment interpreter (formerly `activation-scripts`) into `common`,
`activate` and `wrapper`, where `activate` and `wrapper` only contain the scripts for either operation.
This was done to avoid individually picking files at build time for the two separate outputs.

Additionally, make the package and directory name consistent with its documentation and usage.
For a while now we have been referring to the activation scripts as the `flox-interpreter`.