refactor(activate): Move common Command up
Construct the activate `Command` with exports and args that are common
to all modes earlier on so that we don't have to maintain three
different versions of it.

This will specifically make it easier to move more of the CLI<>Bash
interface from environment variables to getopt arguments to prevent
accidental leaking between nested activations.